Organize app, added new logo,
