+++ src/App.jsx
@@ -1,5 +1,5 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from './assets/Package_logo.png';
+import './css/App.css';
 
 function App() {
   return (
